scripts/cluster-install/requirements.txt
openstacksdk==0.13.0
paramiko==1.18.3
PyYAML==3.10
decorator==4.4.0